+++ cheesylib/src/main/java/com/team254/lib/drivers/TalonSRXFactory.java
@@ -20,8 +20,8 @@ public class TalonSRXFactory {
         public boolean ENABLE_CURRENT_LIMIT = false;
         public boolean ENABLE_SOFT_LIMIT = false;
         public boolean ENABLE_LIMIT_SWITCH = false;
-        public int FORWARD_SOFT_LIMIT = 0;
-        public int REVERSE_SOFT_LIMIT = 0;
+        public double FORWARD_SOFT_LIMIT = 0;
+        public double REVERSE_SOFT_LIMIT = 0;
 
         public boolean INVERTED = false;
         public boolean SENSOR_PHASE = false;
